move boostd-data command into extern/boostd-data (#1093)
* refactor: move boostd-data command into extern/boostd-data

* refactor: move shared code into extern/boostd-data (#1095)
