+++ cli/cmd.go
@@ -1,13 +1,13 @@
 package cli
 
 import (
-	"github.com/filecoin-project/boost/cli/ctxutil"
 	cliutil "github.com/filecoin-project/boost/cli/util"
+	scliutil "github.com/filecoin-project/boostd-data/shared/cliutil"
 	lcliutil "github.com/filecoin-project/lotus/cli/util"
 )
 
 var GetBoostAPI = cliutil.GetBoostAPI
 var GetFullNodeAPI = lcliutil.GetFullNodeAPI
 
-var ReqContext = ctxutil.ReqContext
-var DaemonContext = ctxutil.DaemonContext
+var ReqContext = scliutil.ReqContext
+var DaemonContext = scliutil.DaemonContext
