+++ cmd/booster-http/run.go
@@ -3,13 +3,13 @@ package main
 import (
 	"context"
 	"fmt"
+	"github.com/filecoin-project/boostd-data/shared/tracing"
 	"net/http"
 	_ "net/http/pprof"
 
 	"github.com/filecoin-project/boost/cmd/lib"
 	"github.com/filecoin-project/boost/markets/dagstore"
 	"github.com/filecoin-project/boost/piecedirectory"
-	"github.com/filecoin-project/boost/tracing"
 	"github.com/filecoin-project/boostd-data/model"
 	"github.com/filecoin-project/dagstore/mount"
 	"github.com/filecoin-project/go-state-types/abi"
