+++ extern/boostd-data/ldb/service.go
@@ -8,8 +8,8 @@ import (
 	"sync"
 	"time"
 
-	"github.com/filecoin-project/boost/tracing"
 	"github.com/filecoin-project/boostd-data/model"
+	"github.com/filecoin-project/boostd-data/shared/tracing"
 	"github.com/filecoin-project/boostd-data/svc/types"
 	"github.com/ipfs/go-cid"
 	"github.com/ipfs/go-datastore"
