+++ piecedirectory/piecedirectory.go
@@ -6,13 +6,13 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
+	"github.com/filecoin-project/boostd-data/shared/tracing"
 	"io"
 	"sync"
 	"time"
 
 	"github.com/filecoin-project/boost/markets/dagstore"
 	"github.com/filecoin-project/boost/piecedirectory/types"
-	"github.com/filecoin-project/boost/tracing"
 	"github.com/filecoin-project/boostd-data/client"
 	"github.com/filecoin-project/boostd-data/model"
 	"github.com/filecoin-project/go-state-types/abi"
